make IP-Adapter generic for SD1 and SDXL
